+++ packages/address-book-controller/src/AddressBookController.test.ts
@@ -1,4 +1,4 @@
-import { AddressBookController } from './AddressBookController';
+import { AddressBookController, AddressType } from './AddressBookController';
 
 describe('AddressBookController', () => {
   it('should set default state', () => {
@@ -9,6 +9,7 @@ describe('AddressBookController', () => {
   it('should add a contact entry', () => {
     const controller = new AddressBookController();
     controller.set('0x32Be343B94f860124dC4fEe278FDCBD38C102D88', 'foo');
+
     expect(controller.state).toStrictEqual({
       addressBook: {
         1: {
@@ -18,6 +19,7 @@ describe('AddressBookController', () => {
             isEns: false,
             memo: '',
             name: 'foo',
+            addressType: undefined,
           },
         },
       },
@@ -31,6 +33,7 @@ describe('AddressBookController', () => {
       'foo',
       '1',
       'account 1',
+      AddressType.externallyOwnedAccounts,
     );
 
     expect(controller.state).toStrictEqual({
@@ -42,6 +45,59 @@ describe('AddressBookController', () => {
             isEns: false,
             memo: 'account 1',
             name: 'foo',
+            addressType: AddressType.externallyOwnedAccounts,
+          },
+        },
+      },
+    });
+  });
+
+  it('should add a contact entry with address type contract accounts', () => {
+    const controller = new AddressBookController();
+    controller.set(
+      '0x32Be343B94f860124dC4fEe278FDCBD38C102D88',
+      'foo',
+      '1',
+      'account 1',
+      AddressType.contractAccounts,
+    );
+
+    expect(controller.state).toStrictEqual({
+      addressBook: {
+        1: {
+          '0x32Be343B94f860124dC4fEe278FDCBD38C102D88': {
+            address: '0x32Be343B94f860124dC4fEe278FDCBD38C102D88',
+            chainId: '1',
+            isEns: false,
+            memo: 'account 1',
+            name: 'foo',
+            addressType: AddressType.contractAccounts,
+          },
+        },
+      },
+    });
+  });
+
+  it('should add a contact entry with address type non accounts', () => {
+    const controller = new AddressBookController();
+    controller.set(
+      '0x32Be343B94f860124dC4fEe278FDCBD38C102D88',
+      'foo',
+      '1',
+      'account 1',
+      AddressType.nonAccounts,
+    );
+
+    expect(controller.state).toStrictEqual({
+      addressBook: {
+        1: {
+          '0x32Be343B94f860124dC4fEe278FDCBD38C102D88': {
+            address: '0x32Be343B94f860124dC4fEe278FDCBD38C102D88',
+            chainId: '1',
+            isEns: false,
+            memo: 'account 1',
+            name: 'foo',
+            addressType: AddressType.nonAccounts,
           },
         },
       },
@@ -73,6 +129,7 @@ describe('AddressBookController', () => {
             isEns: false,
             memo: 'account 2',
             name: 'foo',
+            addressType: undefined,
           },
         },
         2: {
@@ -82,6 +139,7 @@ describe('AddressBookController', () => {
             isEns: false,
             memo: 'account 2',
             name: 'foo',
+            addressType: undefined,
           },
         },
       },
@@ -91,7 +149,9 @@ describe('AddressBookController', () => {
   it('should update a contact entry', () => {
     const controller = new AddressBookController();
     controller.set('0x32Be343B94f860124dC4fEe278FDCBD38C102D88', 'foo');
+
     controller.set('0x32Be343B94f860124dC4fEe278FDCBD38C102D88', 'bar');
+
     expect(controller.state).toStrictEqual({
       addressBook: {
         1: {
@@ -101,6 +161,7 @@ describe('AddressBookController', () => {
             isEns: false,
             memo: '',
             name: 'bar',
+            addressType: undefined,
           },
         },
       },
@@ -109,7 +170,7 @@ describe('AddressBookController', () => {
 
   it('should not add invalid contact entry', () => {
     const controller = new AddressBookController();
-    controller.set('0x01', 'foo');
+    controller.set('0x01', 'foo', AddressType.externallyOwnedAccounts);
     expect(controller.state).toStrictEqual({ addressBook: {} });
   });
 
@@ -124,6 +185,7 @@ describe('AddressBookController', () => {
   it('should remove only one contact entry', () => {
     const controller = new AddressBookController();
     controller.set('0x32Be343B94f860124dC4fEe278FDCBD38C102D88', 'foo');
+
     controller.set('0xc38bf1ad06ef69f0c04e29dbeb4152b4175f0a8d', 'bar');
     controller.delete('1', '0xc38bf1ad06ef69f0c04e29dbeb4152b4175f0a8d');
 
@@ -136,6 +198,7 @@ describe('AddressBookController', () => {
             isEns: false,
             memo: '',
             name: 'foo',
+            addressType: undefined,
           },
         },
       },
@@ -145,6 +208,7 @@ describe('AddressBookController', () => {
   it('should add two contact entries with the same chainId', () => {
     const controller = new AddressBookController();
     controller.set('0x32Be343B94f860124dC4fEe278FDCBD38C102D88', 'foo');
+
     controller.set('0xc38bf1ad06ef69f0c04e29dbeb4152b4175f0a8d', 'bar');
 
     expect(controller.state).toStrictEqual({
@@ -156,6 +220,7 @@ describe('AddressBookController', () => {
             isEns: false,
             memo: '',
             name: 'foo',
+            addressType: undefined,
           },
           '0xC38bF1aD06ef69F0c04E29DBeB4152B4175f0A8D': {
             address: '0xC38bF1aD06ef69F0c04E29DBeB4152B4175f0A8D',
@@ -163,6 +228,7 @@ describe('AddressBookController', () => {
             isEns: false,
             memo: '',
             name: 'bar',
+            addressType: undefined,
           },
         },
       },
@@ -185,6 +251,7 @@ describe('AddressBookController', () => {
             isEns: true,
             memo: '',
             name: 'metamask.eth',
+            addressType: undefined,
           },
         },
       },
@@ -194,6 +261,7 @@ describe('AddressBookController', () => {
   it('should clear all contact entries', () => {
     const controller = new AddressBookController();
     controller.set('0x32Be343B94f860124dC4fEe278FDCBD38C102D88', 'foo');
+
     controller.set('0xc38bf1ad06ef69f0c04e29dbeb4152b4175f0a8d', 'bar');
     controller.clear();
     expect(controller.state).toStrictEqual({ addressBook: {} });
@@ -208,12 +276,15 @@ describe('AddressBookController', () => {
 
   it('should return false to indicate an address book entry has NOT been added', () => {
     const controller = new AddressBookController();
-    expect(controller.set('0x00', 'foo')).toStrictEqual(false);
+    expect(
+      controller.set('0x00', 'foo', AddressType.externallyOwnedAccounts),
+    ).toStrictEqual(false);
   });
 
   it('should return true to indicate an address book entry has been deleted', () => {
     const controller = new AddressBookController();
     controller.set('0x32Be343B94f860124dC4fEe278FDCBD38C102D88', 'foo');
+
     expect(
       controller.delete('1', '0x32Be343B94f860124dC4fEe278FDCBD38C102D88'),
     ).toStrictEqual(true);
@@ -228,6 +299,7 @@ describe('AddressBookController', () => {
   it('should normalize addresses so adding and removing entries work across casings', () => {
     const controller = new AddressBookController();
     controller.set('0x32Be343B94f860124dC4fEe278FDCBD38C102D88', 'foo');
+
     controller.set('0xc38bf1ad06ef69f0c04e29dbeb4152b4175f0a8d', 'bar');
 
     controller.delete('1', '0xC38BF1AD06EF69F0C04E29DBEB4152B4175F0A8D');
@@ -240,6 +312,7 @@ describe('AddressBookController', () => {
             isEns: false,
             memo: '',
             name: 'foo',
+            addressType: undefined,
           },
         },
       },