+++ packages/assets-controllers/src/TokenListController.test.ts
@@ -26,7 +26,7 @@ const sampleMainnetTokenList = [
     occurrences: 11,
     name: 'Synthetix',
     iconUrl:
-      'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f.png',
+      'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f.png',
     aggregators: [
       'Aave',
       'Bancor',
@@ -49,7 +49,7 @@ const sampleMainnetTokenList = [
     occurrences: 11,
     name: 'Chainlink',
     iconUrl:
-      'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
+      'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
     aggregators: [
       'Aave',
       'Bancor',
@@ -71,7 +71,7 @@ const sampleMainnetTokenList = [
     occurrences: 11,
     name: 'Bancor',
     iconUrl:
-      'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x1f573d6fb3f13d689ff844b4ce37794d79a7ff1c.png',
+      'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x1f573d6fb3f13d689ff844b4ce37794d79a7ff1c.png',
     aggregators: [
       'Bancor',
       'CMC',
@@ -102,7 +102,7 @@ const sampleWithDuplicateSymbols = [
     occurrences: 11,
     name: 'Bancor',
     iconUrl:
-      'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x1f573d6fb3f13d689ff844b4ce37794d79a7ff1c.png',
+      'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x1f573d6fb3f13d689ff844b4ce37794d79a7ff1c.png',
     aggregators: [
       'Bancor',
       'CMC',
@@ -131,7 +131,7 @@ const sampleWithLessThan3OccurencesResponse = [
     occurrences: 2,
     name: 'Synthetix',
     iconUrl:
-      'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f.png',
+      'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f.png',
     aggregators: [
       'Aave',
       'Bancor',
@@ -154,7 +154,7 @@ const sampleWithLessThan3OccurencesResponse = [
     occurrences: 11,
     name: 'Chainlink',
     iconUrl:
-      'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
+      'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
     aggregators: [
       'Aave',
       'Bancor',
@@ -194,7 +194,7 @@ const sampleBinanceTokenList = [
       'Paraswap',
     ],
     iconUrl:
-      'https://static.metaswap.codefi.network/api/v1/tokenIcons/56/0x7083609fce4d1d8dc0c979aab8c869ea2c873402.png',
+      'https://static.metafi.codefi.network/api/v1/tokenIcons/56/0x7083609fce4d1d8dc0c979aab8c869ea2c873402.png',
   },
   {
     address: '0x1af3f329e8be154074d8769d1ffa4ee058b1dbc3',
@@ -211,7 +211,7 @@ const sampleBinanceTokenList = [
       'Paraswap',
     ],
     iconUrl:
-      'https://static.metaswap.codefi.network/api/v1/tokenIcons/56/0x1af3f329e8be154074d8769d1ffa4ee058b1dbc3.png',
+      'https://static.metafi.codefi.network/api/v1/tokenIcons/56/0x1af3f329e8be154074d8769d1ffa4ee058b1dbc3.png',
   },
 ];
 const sampleSingleChainState = {
@@ -223,7 +223,7 @@ const sampleSingleChainState = {
       occurrences: 11,
       name: 'Synthetix',
       iconUrl:
-        'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f.png',
+        'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f.png',
       aggregators: [
         'Aave',
         'Bancor',
@@ -246,7 +246,7 @@ const sampleSingleChainState = {
       occurrences: 11,
       name: 'Chainlink',
       iconUrl:
-        'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
+        'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
       aggregators: [
         'Aave',
         'Bancor',
@@ -268,7 +268,7 @@ const sampleSingleChainState = {
       occurrences: 11,
       name: 'Bancor',
       iconUrl:
-        'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x1f573d6fb3f13d689ff844b4ce37794d79a7ff1c.png',
+        'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x1f573d6fb3f13d689ff844b4ce37794d79a7ff1c.png',
       aggregators: [
         'Bancor',
         'CMC',
@@ -314,7 +314,7 @@ const sampleTwoChainState = {
         'Paraswap',
       ],
       iconUrl:
-        'https://static.metaswap.codefi.network/api/v1/tokenIcons/56/0x7083609fce4d1d8dc0c979aab8c869ea2c873402.png',
+        'https://static.metafi.codefi.network/api/v1/tokenIcons/56/0x7083609fce4d1d8dc0c979aab8c869ea2c873402.png',
     },
     '0x1af3f329e8be154074d8769d1ffa4ee058b1dbc3': {
       address: '0x1af3f329e8be154074d8769d1ffa4ee058b1dbc3',
@@ -331,7 +331,7 @@ const sampleTwoChainState = {
         'Paraswap',
       ],
       iconUrl:
-        'https://static.metaswap.codefi.network/api/v1/tokenIcons/56/0x1af3f329e8be154074d8769d1ffa4ee058b1dbc3.png',
+        'https://static.metafi.codefi.network/api/v1/tokenIcons/56/0x1af3f329e8be154074d8769d1ffa4ee058b1dbc3.png',
     },
   },
   tokensChainsCache: {
@@ -355,7 +355,7 @@ const existingState = {
       occurrences: 11,
       name: 'Chainlink',
       iconUrl:
-        'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
+        'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
       aggregators: [
         'Aave',
         'Bancor',
@@ -389,7 +389,7 @@ const outdatedExistingState = {
       occurrences: 11,
       name: 'Chainlink',
       iconUrl:
-        'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
+        'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
       aggregators: [
         'Aave',
         'Bancor',
@@ -423,7 +423,7 @@ const expiredCacheExistingState: TokenListState = {
       occurrences: 9,
       name: 'Chainlink',
       iconUrl:
-        'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
+        'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
       aggregators: [
         'Aave',
         'Bancor',
@@ -450,7 +450,7 @@ const expiredCacheExistingState: TokenListState = {
           occurrences: 11,
           name: 'Chainlink',
           iconUrl:
-            'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
+            'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
           aggregators: [
             'Aave',
             'Bancor',
@@ -559,7 +559,7 @@ describe('TokenListController', () => {
           occurrences: 11,
           name: 'Chainlink',
           iconUrl:
-            'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
+            'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x514910771af9ca656af840dff83e8264ecf986ca.png',
           aggregators: [
             'Aave',
             'Bancor',
@@ -946,7 +946,7 @@ describe('TokenListController', () => {
         occurrences: 11,
         name: 'Bancor',
         iconUrl:
-          'https://static.metaswap.codefi.network/api/v1/tokenIcons/1/0x1f573d6fb3f13d689ff844b4ce37794d79a7ff1c.png',
+          'https://static.metafi.codefi.network/api/v1/tokenIcons/1/0x1f573d6fb3f13d689ff844b4ce37794d79a7ff1c.png',
         aggregators: [
           'Bancor',
           'CMC',
@@ -1069,8 +1069,8 @@ describe('TokenListController', () => {
     nock(TOKEN_END_POINT_API)
       .get(`/tokens/${NetworksChainId.mainnet}`)
       .reply(200, sampleMainnetTokenList)
-      .get(`/tokens/${NetworksChainId.ropsten}`)
-      .reply(200, { error: 'ChainId 3 is not supported' })
+      .get(`/tokens/${NetworksChainId.goerli}`)
+      .reply(200, { error: 'ChainId 5 is not supported' })
       .get(`/tokens/56`)
       .reply(200, sampleBinanceTokenList)
       .persist();
@@ -1097,7 +1097,7 @@ describe('TokenListController', () => {
       sampleTwoChainState.tokensChainsCache[NetworksChainId.mainnet].data,
     );
 
-    network.setProviderType('ropsten');
+    network.setProviderType('goerli');
 
     await new Promise<void>((resolve) => setTimeout(() => resolve(), 500));
 
@@ -1157,8 +1157,8 @@ describe('TokenListController', () => {
     nock(TOKEN_END_POINT_API)
       .get(`/tokens/${NetworksChainId.mainnet}`)
       .reply(200, sampleMainnetTokenList)
-      .get(`/tokens/${NetworksChainId.ropsten}`)
-      .reply(200, { error: 'ChainId 3 is not supported' })
+      .get(`/tokens/${NetworksChainId.goerli}`)
+      .reply(200, { error: 'ChainId 5 is not supported' })
       .get(`/tokens/56`)
       .reply(200, sampleBinanceTokenList)
       .persist();
@@ -1167,7 +1167,7 @@ describe('TokenListController', () => {
     const { network } = setupNetworkController(controllerMessenger);
     const messenger = getRestrictedMessenger(controllerMessenger);
     const controller = new TokenListController({
-      chainId: NetworksChainId.ropsten,
+      chainId: NetworksChainId.goerli,
       preventPollingOnNetworkRestart: true,
       messenger,
       interval: 100,