+++ packages/network-controller/src/NetworkController.ts
@@ -73,6 +73,10 @@ export type NetworkControllerProviderConfigChangeEvent = {
   payload: [ProviderConfig];
 };
 
+export type NetworkControllerEvents =
+  | NetworkControllerStateChangeEvent
+  | NetworkControllerProviderConfigChangeEvent;
+
 export type NetworkControllerGetProviderConfigAction = {
   type: `NetworkController:getProviderConfig`;
   handler: () => ProviderConfig;
@@ -82,6 +86,10 @@ export type NetworkControllerGetEthQueryAction = {
   type: `NetworkController:getEthQuery`;
   handler: () => EthQuery;
 };
+
+export type NetworkControllerActions =
+  | NetworkControllerGetProviderConfigAction
+  | NetworkControllerGetEthQueryAction;
 
 export type NetworkControllerMessenger = RestrictedControllerMessenger<
   typeof name,
@@ -98,7 +106,7 @@ export type NetworkControllerOptions = {
   state?: Partial<NetworkState>;
 };
 
-const defaultState: NetworkState = {
+export const defaultState: NetworkState = {
   network: 'loading',
   isCustomNetwork: false,
   providerConfig: { type: MAINNET, chainId: NetworksChainId.mainnet },
@@ -173,11 +181,9 @@ export class NetworkController extends BaseControllerV2<
     });
 
     switch (type) {
-      case 'kovan':
       case MAINNET:
-      case 'rinkeby':
       case 'goerli':
-      case 'ropsten':
+      case 'sepolia':
         this.setupInfuraProvider(type);
         break;
       case 'localhost':
@@ -230,10 +236,8 @@ export class NetworkController extends BaseControllerV2<
   private getIsCustomNetwork(chainId?: string) {
     return (
       chainId !== NetworksChainId.mainnet &&
-      chainId !== NetworksChainId.kovan &&
-      chainId !== NetworksChainId.rinkeby &&
       chainId !== NetworksChainId.goerli &&
-      chainId !== NetworksChainId.ropsten &&
+      chainId !== NetworksChainId.sepolia &&
       chainId !== NetworksChainId.localhost
     );
   }
@@ -290,7 +294,9 @@ export class NetworkController extends BaseControllerV2<
     const { type, rpcTarget, chainId, ticker, nickname } =
       this.state.providerConfig;
     this.initializeProvider(type, rpcTarget, chainId, ticker, nickname);
-    this.registerProvider();
+    if (this.provider !== undefined) {
+      this.registerProvider();
+    }
     this.lookupNetwork();
   }
 
@@ -298,38 +304,53 @@ export class NetworkController extends BaseControllerV2<
     throw new Error('Property only used for setting');
   }
 
+  async #getNetworkId(): Promise<string> {
+    return await new Promise((resolve, reject) => {
+      this.ethQuery.sendAsync(
+        { method: 'net_version' },
+        (error: Error, result: string) => {
+          if (error) {
+            reject(error);
+          } else {
+            resolve(result);
+          }
+        },
+      );
+    });
+  }
+
   /**
    * Refreshes the current network code.
    */
   async lookupNetwork() {
-    /* istanbul ignore if */
     if (!this.ethQuery || !this.ethQuery.sendAsync) {
       return;
     }
     const releaseLock = await this.mutex.acquire();
-    this.ethQuery.sendAsync(
-      { method: 'net_version' },
-      (error: Error, network: string) => {
-        try {
-          if (this.state.network === network) {
-            return;
-          }
 
-          this.update((state) => {
-            state.network = error
-              ? /* istanbul ignore next*/ 'loading'
-              : network;
-          });
-
-          this.messagingSystem.publish(
-            `NetworkController:providerConfigChange`,
-            this.state.providerConfig,
-          );
-        } finally {
-          releaseLock();
+    try {
+      try {
+        const networkId = await this.#getNetworkId();
+        if (this.state.network === networkId) {
+          return;
         }
-      },
-    );
+
+        this.update((state) => {
+          state.network = networkId;
+        });
+      } catch (_error) {
+        this.update((state) => {
+          state.network = 'loading';
+        });
+      }
+
+      this.messagingSystem.publish(
+        `NetworkController:providerConfigChange`,
+        this.state.providerConfig,
+      );
+    } finally {
+      releaseLock();
+    }
   }
 
   /**